+++ pkg/backend/httpstate/backend.go
@@ -118,7 +118,9 @@ type Backend interface {
 	Search(
 		ctx context.Context, orgName string, queryParams *apitype.PulumiQueryRequest,
 	) (*apitype.ResourceSearchResponse, error)
-	NaturalLanguageSearch(ctx context.Context, orgName string, query string) (*apitype.ResourceSearchResponse, error)
+	NaturalLanguageSearch(
+		ctx context.Context, orgName string, query string,
+	) (*apitype.ResourceSearchResponse, error)
 }
 
 type cloudBackend struct {
@@ -1027,7 +1029,9 @@ func (b *cloudBackend) Query(ctx context.Context, op backend.QueryOperation) res
 func (b *cloudBackend) Search(
 	ctx context.Context, orgName string, queryParams *apitype.PulumiQueryRequest,
 ) (*apitype.ResourceSearchResponse, error) {
-	return b.Client().GetSearchQueryResults(ctx, orgName, queryParams)
+	results, err := b.Client().GetSearchQueryResults(ctx, orgName, queryParams, b.CloudConsoleURL())
+	results.Query = queryParams.Query
+	return results, err
 }
 
 func (b *cloudBackend) NaturalLanguageSearch(
@@ -1038,7 +1042,8 @@ func (b *cloudBackend) NaturalLanguageSearch(
 		return nil, err
 	}
 	requestBody := apitype.PulumiQueryRequest{Query: parsedResults.Query}
-	results, err := b.Client().GetSearchQueryResults(ctx, orgName, &requestBody)
+	results, err := b.Client().GetSearchQueryResults(ctx, orgName, &requestBody, b.CloudConsoleURL())
+	results.Query = parsedResults.Query
 	if err != nil {
 		return nil, err
 	}