+++ pkg/cmd/pulumi/org_search_test.go
@@ -17,6 +17,7 @@ package main
 import (
 	"bytes"
 	"context"
+	"fmt"
 	"testing"
 
 	"github.com/pulumi/pulumi/pkg/v3/backend"
@@ -38,29 +39,37 @@ func TestSearch_cmd(t *testing.T) {
 	pack := "pack1"
 	mod := "mod1"
 	modified := "2023-01-01T00:00:00.000Z"
-	cmd := searchCmd{
-		Stdout: &buff,
-		currentBackend: func(context.Context, *workspace.Project, display.Options) (backend.Backend, error) {
-			return &stubHTTPBackend{
-				SearchF: func(context.Context, string, *apitype.PulumiQueryRequest) (*apitype.ResourceSearchResponse, error) {
-					return &apitype.ResourceSearchResponse{
-						Resources: []apitype.ResourceResult{
-							{
-								Name:     &name,
-								Type:     &typ,
-								Program:  &program,
-								Stack:    &stack,
-								Package:  &pack,
-								Module:   &mod,
-								Modified: &modified,
+	searchURL := "https://app.pulumi.com/pulumi/resources?foo=bar"
+	total := int64(132)
+	orgName := "org1"
+	cmd := orgSearchCmd{
+		searchCmd: searchCmd{
+			orgName: orgName,
+			Stdout:  &buff,
+			currentBackend: func(context.Context, *workspace.Project, display.Options) (backend.Backend, error) {
+				return &stubHTTPBackend{
+					SearchF: func(context.Context, string, *apitype.PulumiQueryRequest) (*apitype.ResourceSearchResponse, error) {
+						return &apitype.ResourceSearchResponse{
+							Resources: []apitype.ResourceResult{
+								{
+									Name:     &name,
+									Type:     &typ,
+									Program:  &program,
+									Stack:    &stack,
+									Package:  &pack,
+									Module:   &mod,
+									Modified: &modified,
+								},
 							},
-						},
-					}, nil
-				},
-				CurrentUserF: func() (string, []string, error) {
-					return "user", []string{"org1", "org2"}, nil
-				},
-			}, nil
+							URL:   searchURL,
+							Total: &total,
+						}, nil
+					},
+					CurrentUserF: func() (string, []string, error) {
+						return "user", []string{"org1", "org2"}, nil
+					},
+				}, nil
+			},
 		},
 	}
 
@@ -70,6 +79,8 @@ func TestSearch_cmd(t *testing.T) {
 	assert.Contains(t, buff.String(), name)
 	assert.Contains(t, buff.String(), typ)
 	assert.Contains(t, buff.String(), program)
+	assert.Contains(t, buff.String(), fmt.Sprintf("Results are also visible in Pulumi Cloud:\n%s", searchURL))
+	assert.Contains(t, buff.String(), fmt.Sprint(total))
 }
 
 type stubHTTPBackend struct {